+++ angular-client/src/app/app.component.html
@@ -1,8 +1,8 @@
 <mat-toolbar color="primary">
     SocialWorker
   <span *ngIf="this.authenticationService.loggedIn">
-    <a class="nav-link" routerLink="profiles">
-      <button mat-button>Profiles</button>
+    <a class="nav-link" routerLink="participants">
+      <button mat-button>Participants</button>
     </a>
     <a class="nav-link" routerLink="resources">
       <button mat-button>Resources</button>
