+++ angular-client/src/app/components/participants/participant-profile/participant-profile.component.html
