+++ angular-client/src/app/components/participants/participant-profile/participant-profile.component.spec.ts
@@ -1,14 +1,14 @@
 import { async, ComponentFixture, TestBed, tick, fakeAsync } from '@angular/core/testing';
 import { ParticipantProfileComponent } from './participant-profile.component';
-import { MaterialsModule } from '../../modules/materials.module';
+import { MaterialsModule } from '../../../modules/materials.module';
 import { ActivatedRoute, convertToParamMap, Router } from '@angular/router';
-import { ParticipantService } from '../../services/participant.service';
-import { MessageService } from '../../services/message.service';
+import { ParticipantService } from '../../../services/participant.service';
+import { MessageService } from '../../../services/message.service';
 import { RouterTestingModule } from '@angular/router/testing';
 import { Location } from '@angular/common';
-import { MockParticipantService } from '../../mocks/MockParticipantService';
-import { AuthenticationService } from '../../services/authentication.service';
-import { MockAuthenticationService } from '../../mocks/MockAuthenticationService';
+import { MockParticipantService } from '../../../mocks/MockParticipantService';
+import { AuthenticationService } from '../../../services/authentication.service';
+import { MockAuthenticationService } from '../../../mocks/MockAuthenticationService';
 
 describe('ParticipantProfileComponent', () => {
   let component: ParticipantProfileComponent;
@@ -21,10 +21,10 @@ describe('ParticipantProfileComponent', () => {
       declarations: [ ParticipantProfileComponent ],
       providers: [
         MessageService,
-        { provide: ParticipantService, useClass: MockParticipantService}, 
-        { provide: AuthenticationService, useClass: MockAuthenticationService }, 
-        { provide: Router, useValue: { navigateByUrl: jasmine.createSpy("navigateByUrl") } },
-        { provide: ActivatedRoute, useValue: { snapshot: { paramMap: convertToParamMap({_id: "frontendtest"}) } } }
+        { provide: ParticipantService, useClass: MockParticipantService},
+        { provide: AuthenticationService, useClass: MockAuthenticationService },
+        { provide: Router, useValue: { navigateByUrl: jasmine.createSpy('navigateByUrl') } },
+        { provide: ActivatedRoute, useValue: { snapshot: { paramMap: convertToParamMap({_id: 'frontendtest'}) } } }
       ]
     });
     location = TestBed.get(Location);
@@ -37,25 +37,25 @@ describe('ParticipantProfileComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it("should display name of the user", () => {
+  it('should display name of the user', () => {
     const compiled = fixture.debugElement.nativeElement;
     expect(compiled.querySelector('h2').textContent).toContain('testuser');
   });
 
-  it("should display user's note and its attachd image", () => {
+  it('should display user\'s note and its attachd image', () => {
     const compiled = fixture.debugElement.nativeElement;
     expect(compiled.querySelector('mat-expansion-panel').textContent).toContain('Note for testing');
     expect(compiled.querySelector('img')).toBeTruthy();
   });
 
-  it("should refresh resources displayed when one gets deleted", fakeAsync(() => {
+  it('should refresh resources displayed when one gets deleted', fakeAsync(() => {
     const compiled = fixture.debugElement.nativeElement;
     const deleteButton = compiled.querySelector('#deleteBtn');
     deleteButton.click();
 
     tick();
     fixture.detectChanges();
-    
+
     expect(compiled.querySelector('div').textContent).toContain('No notes to display.');
   }));
